==> Media+Publication/SDSHL/SDSHL/notebooks/3-PoC/05.5-embedding_BERT.ipynb <==
!nvidia-smi

#htop 

#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed' #Change this when working on Nimblebox
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models\BERT'
resultsfolder = r'D:\18-DS\github\SDSHL\data\results'

# vectorfolder_bert = modelfolder + r'\BERT'
# vectorfolder_ft   = modelfolder + r'\fasttext_wiki.hi'
# vectorfolder_standford = modelfolder + r'\POS-Tagger-Hindi'

filepath_train =datafolder_p+ r'\2-train.csv' #Change this when working on Nimblebox
filepath_test  =datafolder_p+ r'\2-test.csv'  #Change this when working on Nimblebox
filepath_fulldata = datafolder_p+ r'\2-Hinglish_Sarcasm_Clean.csv'

import pandas as pd
import numpy as np
import sys
import random as rn

import sklearn.metrics as metrics
from sklearn.naive_bayes import GaussianNB
#import sklearn.svm as svm
from sklearn.linear_model import LogisticRegression             
from sklearn.pipeline import Pipeline

from keras.preprocessing.sequence import pad_sequences

import torch
from torch import nn
from pytorch_pretrained_bert import BertModel
from pytorch_pretrained_bert import BertTokenizer

from torch.utils.data import TensorDataset, DataLoader, RandomSampler, SequentialSampler
from torch.optim import Adam
from torch.nn.utils import clip_grad_norm_
#from torchnlp.datasets import imdb_dataset

from IPython.display import clear_output

df_train = pd.read_csv(filepath_train, sep="\t")
df_test  = pd.read_csv(filepath_test, sep="\t")
df_full  = pd.read_csv(filepath_fulldata, sep="\t", index_col="ID")

X_train =  df_train['sentence'] 
y_train = df_train['label']

X_test =  df_test['sentence'] 
y_test = df_test['label']



prediction={}

## Prepare Tensor Dataset of Train and Test Dataset

X_train = list(X_train)
y_train = list(y_train)

X_test = list(X_test)
y_test = list(y_test)

len(X_train), len(y_train), len(X_test), len(y_test)

#!pip install transformers

## Common Functions

from typing import Callable, List, Optional, Tuple

import pandas as pd
from sklearn.base import TransformerMixin, BaseEstimator
import torch


class BertTransformer(BaseEstimator, TransformerMixin):
    def __init__(
            self,
            bert_tokenizer,
            bert_model,
            max_length: int = 60,
            embedding_func: Optional[Callable[[torch.tensor], torch.tensor]] = None,
    ):
        self.tokenizer = bert_tokenizer
        self.model = bert_model
        self.model.eval()
        self.max_length = max_length
        self.embedding_func = embedding_func

        if self.embedding_func is None:
            self.embedding_func = lambda x: x[0][:, 0, :].squeeze()

    def _tokenize(self, text: str) -> Tuple[torch.tensor, torch.tensor]:
        # Tokenize the text with the provided tokenizer
        tokenized_text = self.tokenizer.encode_plus(text,
                                                    add_special_tokens=True,
                                                    max_length=self.max_length,
                                                    truncation=True
                                                    )["input_ids"]

        # Create an attention mask telling BERT to use all words
        attention_mask = [1] * len(tokenized_text)

        # bert takes in a batch so we need to unsqueeze the rows
        return (
            torch.tensor(tokenized_text).unsqueeze(0),
            torch.tensor(attention_mask).unsqueeze(0),
        )

    def _tokenize_and_predict(self, text: str) -> torch.tensor:
        tokenized, attention_mask = self._tokenize(text)

        embeddings = self.model(tokenized, attention_mask)
        return self.embedding_func(embeddings)

    def transform(self, text: List[str]):
        if isinstance(text, pd.Series):
            text = text.tolist()

        with torch.no_grad():
            return torch.stack([self._tokenize_and_predict(string) for string in text])

    def fit(self, X, y=None):
        """No fitting necessary so we just return ourselves"""
        return self

def pred_for_test(model):
    y_pred=[]
    for i in range(len(X_test)):
        p=model.predict(X_test.iloc[i])
        x1 = int(p[0][-1][-1])
        x2 = p[1][0]
        if x1==0: x2= 1-x2
        y_pred.append( x2 )
    return y_pred    


def print_metrics(y_pred):
    threshold=0.5
    y_pred1=[]
    for i in y_pred :
        if i>threshold:
            y_pred1.append(1)
        else:
            y_pred1.append(0)

    acc = np.round( metrics.accuracy_score(y_test,y_pred1), 2)
    recall = np.round( metrics.recall_score(y_test,y_pred1), 2)
    precision = np.round( metrics.precision_score(y_test,y_pred1), 2)
    f1 = np.round( metrics.f1_score(y_test,y_pred1), 2)
    roc = np.round( metrics.roc_auc_score(y_test,y_pred1), 2)

    print("Accuracy : ", acc )
    print("Recall   : ", recall )
    print("Precision: ", precision )
    print("F1       : ", f1 )
    print("ROC      : ", roc )

# <font color=red> BERT Multilanguage - Task Transfer - Pytorch</font>
<font color=red> pytoch works only on Nimblebox. Because it need GPU and my machine does not have that. So DON'T RUN it Here</font>

#Inspiration from
#https://towardsdatascience.com/bert-to-the-rescue-17671379687f
#https://github.com/shudima/notebooks/blob/master/BERT_to_the_rescue.ipynb

def get_bert_tokens():
    #This will download Pretrained Multilingual Model and create a Model in Memory
    bert_tokenizer = BertTokenizer.from_pretrained('bert-base-multilingual-uncased')

    train_tokens = list(map(lambda t: ['[CLS]'] + bert_tokenizer.tokenize(t)[:255] + ['[SEP]'], X_train))
    test_tokens = list(map(lambda t: ['[CLS]'] + bert_tokenizer.tokenize(t)[:255] + ['[SEP]'], X_test))

    print ("Tokens:", len(train_tokens), len(test_tokens) )

    train_tokens_ids = pad_sequences(list(map(bert_tokenizer.convert_tokens_to_ids, train_tokens)), maxlen=256, truncating="post", padding="post", dtype="int")
    test_tokens_ids = pad_sequences(list(map(bert_tokenizer.convert_tokens_to_ids, test_tokens)), maxlen=256, truncating="post", padding="post", dtype="int")

    print ("Token Ids:",train_tokens_ids.shape, test_tokens_ids.shape)

    train_masks = [[float(i > 0) for i in ii] for ii in train_tokens_ids]
    test_masks = [[float(i > 0) for i in ii] for ii in test_tokens_ids]

    print ("Masks: ", len(train_masks), len(test_masks))
    return train_tokens, test_tokens, train_tokens_ids, test_tokens_ids, train_masks, test_masks


BATCH_SIZE = 2
EPOCHS = 10
def get_tensor_dataset(train_tokens_ids,test_tokens_ids,y_train,y_test, train_masks, test_masks):
    train_tokens_tensor = torch.tensor(train_tokens_ids)
    train_y_tensor = torch.tensor(y_train).float()

    test_tokens_tensor = torch.tensor(test_tokens_ids)
    test_y_tensor = torch.tensor(y_test).float()

    train_masks_tensor = torch.tensor(train_masks)
    test_masks_tensor = torch.tensor(test_masks)

    print ( str(torch.cuda.memory_allocated(torch.cuda.current_device())/1000000 ) + 'M')

    train_dataset = TensorDataset(train_tokens_tensor, train_masks_tensor, train_y_tensor)
    #train_sampler = RandomSampler(train_dataset)
    #train_dataloader = DataLoader(train_dataset, sampler=train_sampler, batch_size=BATCH_SIZE)
    train_dataloader = DataLoader(train_dataset,  batch_size=BATCH_SIZE)

    test_dataset = TensorDataset(test_tokens_tensor, test_masks_tensor, test_y_tensor)
    #test_sampler = SequentialSampler(test_dataset)
    test_dataloader = DataLoader(test_dataset, batch_size=BATCH_SIZE)
    
    return train_dataloader, test_dataloader


train_tokens, test_tokens, train_tokens_ids, test_tokens_ids, train_masks, test_masks = get_bert_tokens()

train_dataloader, test_dataloader  = \
get_tensor_dataset(train_tokens_ids,test_tokens_ids,y_train,y_test, train_masks, test_masks)

### Develop BERT Multilanguage Model

#This is required for multilingual bert

class BertBinaryClassifier(nn.Module):
    def __init__(self, dropout=0.1):
        super(BertBinaryClassifier, self).__init__()
        self.bert = BertModel.from_pretrained( 'bert-base-multilingual-uncased')  

        self.dropout = nn.Dropout(dropout)
        self.linear = nn.Linear(768, 1)
        self.sigmoid = nn.Sigmoid()
    
    def forward(self, tokens, masks=None):
        _, pooled_output = self.bert(tokens, attention_mask=masks, output_all_encoded_layers=False)
        #_, pooled_output = self.bert(tokens, attention_mask=masks)
        dropout_output = self.dropout(pooled_output)
        linear_output = self.linear(dropout_output)
        proba = self.sigmoid(linear_output)
        return proba

bert_clf = BertBinaryClassifier()
bert_clf = bert_clf.cuda()

param_optimizer = list(bert_clf.sigmoid.named_parameters()) 
optimizer_grouped_parameters = [{"params": [p for n, p in param_optimizer]}]

optimizer = Adam(bert_clf.parameters(), lr=3e-6)

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
device

def train_bert():
    for epoch_num in range(EPOCHS):
        bert_clf.train()
        train_loss = 0
        for step_num, batch_data in enumerate(train_dataloader):
            token_ids, masks, labels = tuple(t.to(device) for t in batch_data)
            #print(str(torch.cuda.memory_allocated(device)/1000000 ) + 'M')
            logits = bert_clf(token_ids, masks)

            loss_func = nn.BCELoss()

            batch_loss = loss_func(logits, labels)
            train_loss += batch_loss.item()


            bert_clf.zero_grad()
            batch_loss.backward()


            clip_grad_norm_(parameters=bert_clf.parameters(), max_norm=1.0)
            optimizer.step()

            clear_output(wait=True)
            print('Epoch: ', epoch_num + 1)
            print("\r" + "{0}/{1} loss: {2} ".format(step_num, len(token_ids) / BATCH_SIZE, train_loss / (step_num + 1)))
            
        return bert_clf

bert_clf = train_bert()
bert_clf = bert_clf.cuda()

### BERT Multilanguage - Evaluation

def evaluate_model():
    bert_clf.eval()
    bert_predicted = []
    all_logits = []
    with torch.no_grad():
        for step_num, batch_data in enumerate(test_dataloader):

            token_ids, masks, labels = tuple(t.to(device) for t in batch_data)

            logits = bert_clf(token_ids, masks)
            loss_func = nn.BCELoss()
            loss = loss_func(logits, labels)
            numpy_logits = logits.cpu().detach().numpy()

            #bert_predicted += list(numpy_logits[:, 0] > 0.5)
            bert_predicted += list(numpy_logits[:, 0])
            
    return bert_predicted

bert_predicted_multi = evaluate_model()

from sklearn.metrics import classification_report
prediction['mBERT_PytorchTT'] = bert_predicted_multi

print(classification_report(y_test, np.array(bert_predicted_multi)>.5))

# <font color=red> mBERT / Multilanguage - Task Transfer - Transformer</font>

from transformers import BertTokenizer as BertTokenizer2, BertModel as BertModel2
bert_model_multi     = BertModel2.from_pretrained(modelfolder + r'\bert-base-multilingual-uncased')
bert_tokenizer_multi = BertTokenizer2.from_pretrained(modelfolder + r'\bert-base-multilingual-uncased')

bert_transformer_multi = BertTransformer(bert_tokenizer_multi, bert_model_multi, max_length=200)
#classifier = svm.LinearSVC(C=1.0, class_weight="balanced")
#classifier = LogisticRegression()
classifier = GaussianNB()

model_pipeline_multi = Pipeline(
    [
        ("vectorizer", bert_transformer_multi),
        ("classifier", classifier),
    ]
)
model_pipeline_multi.fit(X_train, y_train)

### mBERT  /Multilanguage - Transformer - Evaluation

y_pred_multi = model_pipeline_multi.predict_proba(X_test)

prediction['mBERT_TransformerTT'] = y_pred_multi[:,1]
print_metrics(y_pred_multi[:,1])

# <font color=red> mBERT (Multilanguage) - Embedding Transfer</font>

#This steps takes approx 5 min.

import torch

mbert_embedding=[]
for i in range(len(df_full)):
    txt = df_full.iloc[i,:]['sentence']
    token_dict = bert_tokenizer_multi.encode_plus(txt, add_special_tokens=False)
    embed = bert_model_multi((torch.tensor(token_dict["input_ids"]).unsqueeze(0)))
    embed = embed[1][0].detach().numpy()
    #print (i, sep='\t')
    mbert_embedding.append(embed)

df_mbert = pd.DataFrame(mbert_embedding, index=df_full.index)
df_mbert['label'] = df_full['label']
df_mbert.head(5)

#### Save BERT Multilangue Embedding

df_mbert.to_csv(datafolder_p+r'\embedding_bert768_mbert_pretrained.csv')

# <font color=red> IndicBERT -Task Transfer - Transformer</font>

from transformers import AutoTokenizer, AutoModel
bert_tokenizer_indic = AutoTokenizer.from_pretrained("ai4bharat/indic-bert")
bert_model_indic = AutoModel.from_pretrained("ai4bharat/indic-bert")


bert_transformer_indic = BertTransformer(bert_tokenizer_indic, bert_model_indic, max_length=200)
#classifier = svm.LinearSVC(C=1.0, class_weight="balanced")
#classifier = LogisticRegression()
classifier = GaussianNB()

model_pipeline_indic = Pipeline(
    [
        ("vectorizer", bert_transformer_indic),
        ("classifier", classifier),
    ]
)
model_pipeline_indic.fit(X_train, y_train)

### IndicBERT  - Evaluation

y_pred_indic = model_pipeline_indic.predict_proba(X_test)

prediction['IndicBERT_TT'] = y_pred_indic[:,1]
print_metrics(y_pred_indic[:,1])

# <font color=red> IndicBERT - Embedding Transfer</font>

#This steps takes approx 5 min.

indicbert_embedding=[]
for i in range(len(df_full)):
    txt = df_full.iloc[i,:]['sentence']
    token_dict = bert_tokenizer_indic.encode_plus(txt, add_special_tokens=False, max_length=200, truncation=True)
    embed = bert_model_indic((torch.tensor(token_dict["input_ids"]).unsqueeze(0)))
    embed = embed[1][0].detach().numpy()
    #print (i, sep='\t')
    indicbert_embedding.append(embed)

df_indicbert = pd.DataFrame(indicbert_embedding, index=df_full.index)
df_indicbert['label'] = df_full['label']
df_indicbert.head(5)

#### Save BERT indicnlp Embedding

df_indicbert.to_csv(datafolder_p+r'\embedding_bert768_indicbert_pretrained.csv')

### <font color=red> Save Task Transfer Prediction Results of mBERT and indicBERT</font>

df_prediction= pd.DataFrame(prediction, index=df_test.ID)
df_prediction.to_csv(resultsfolder+ r'\model_predictions_BERT.csv')
df_prediction
