==> Media+Publication/SDSHL/SDSHL/notebooks/3-PoC/09-Evaluation-All-Models.ipynb <==
import pandas as pd
import numpy as np
import sklearn.metrics as metrics

#paths required
datafolder_p = r'D:\18-DS\github\SDSHL\data\processed'
datafolder_e = r'D:\18-DS\github\SDSHL\data\external'
datafolder_i = r'D:\18-DS\github\SDSHL\data\internal'
modelfolder  = r'D:\18-DS\data\models'
resultsfolder = r'D:\18-DS\github\SDSHL\data\results'

file_train = datafolder_p + r'\2-train.csv'
file_test = datafolder_p + r'\2-test.csv'
file_org  = datafolder_p + r'\1-Hinglish_SarcasmCSV.csv'

#uncomment variable below to calculate metrics to include prediction of combined embedding
#This file is created after selecting best embedding and combining that with lexical features
#That is possible only when we run all the models of 09.3-Calssification-All-in-One notebook & 
#combined the prediction results of fasttext, bert, cnn
#prediction_file = r'\model_predictions_combined.csv'

model_with_data_list=['Normal','RandomOverSampler','Smote','AdaSyn']
option=0

# Combine Predictions of All the Models

def combine_all_prediction(pred_mainfilename):
    #Reading predictions of classifcal models
    df_prediction  = pd.read_csv( resultsfolder + pred_mainfilename ,index_col="ID" )

    #FastText Prediction using Pretrained Model with Task Finetuning
    df_prediction_fasttext = pd.read_csv( resultsfolder + r'\model_predictions_fasttext.csv', index_col="ID")

    #BERT Prediction using Pretrained Model with Task Finetuning
    df_prediction_BERT = pd.read_csv( resultsfolder + r'\model_predictions_BERT.csv', index_col="ID")

    #Neural Network Prediction
    df_prediction_CNN      = pd.read_csv( resultsfolder + r'\model_predictions_NN.csv', index_col="ID")

    #Combine All Predictions
    df_prediction = pd.concat( [df_prediction,df_prediction_fasttext, df_prediction_BERT, \
                                df_prediction_CNN], axis=1, ignore_index=False )

    #df_prediction = df_prediction.drop(columns=['label'])
    #df_prediction['label'] = label

    #Identify whether sentence is twitter or non-twitter. Also add label to the prediction
    df_org = pd.read_csv(file_org, sep='\t', index_col='ID')
    df_val = pd.read_csv(file_test, sep='\t', index_col='ID')
    idx_val = df_val.index

    #read only those records from original file which has ID equal to Validation dataset and get their label & Twitter flag
    df_prediction['label'] = df_org.loc[idx_val]['label']
    df_prediction['twitter'] = df_org.loc[idx_val]['twitter']
    return df_prediction

df_prediction = combine_all_prediction(r'\model_predictions_Normal.csv')

y_act          = df_prediction['label']

# Calulate Metrics using Predictions done in earlier step

threshold= 0.5

def calculate_metrics(y_act_, df_prediction_,):
    tot_models=df_prediction_.shape[1]-2
    #print ('Total Modles {tot_mod}'.format(tot_mod= tot_models))
    results = np.zeros( (tot_models,9) )

    final_label = lambda x: 1 if x>threshold else 0

    #y_act= df_prediction1.iloc[:, tot_models]
    model_name=[]
    for i in range(tot_models):

        #print (df_prediction_.columns[i], len(y_act_), len(df_prediction_.iloc[:,i]), i )

        model_name.append( df_prediction_.columns[i])

        #print (y_act_)
        results[i][0] = np.round( metrics.roc_auc_score(y_act_,df_prediction_.iloc[:,i]), 2)

        y_pre= df_prediction_.iloc[:,i].apply(final_label)

        results[i][1] = np.round( metrics.accuracy_score(y_act_,y_pre), 2)
        results[i][2] = np.round( metrics.recall_score(y_act_,y_pre), 2)
        results[i][3] = np.round( metrics.precision_score(y_act_,y_pre), 2)
        results[i][4] = np.round( metrics.f1_score(y_act_,y_pre), 2)
        
        cm= np.round( metrics.confusion_matrix(y_act_,y_pre), 2)
        results[i][5] = cm[0,0]
        results[i][6] = cm[0,1]
        results[i][7] = cm[1,0]
        results[i][8] = cm[1,1]
        #print (np.array(cm).reshape(-1))

        #results[i][4] = np.round( metrics.roc_auc_score(y_act,y_pre), 2)

    results1 = pd.DataFrame(results,columns=['AUC','Accuracy','Recall','Precision','F1','TN','FP','FN','TP'], index=model_name)

    return results1

results1 = calculate_metrics(y_act, df_prediction)

pd.set_option( "display.html.border" , 1 )
pd.set_option("display.max_rows",110)

print ("Number of Models {modnum}".format( modnum = len(results1)))

#Results for all Records
results1 = calculate_metrics(y_act, df_prediction)
results1
results1.sort_values("Accuracy",ascending=False)

# Combine Model Prediction

# Here we know the best Embedding. 
# Combine this embedding with Lexical Features in "08.1-Classification-All-in-One-Shot" - At the end of file.
# Create a combined model and generate prediction results
# Check Here Whether Combined model is any better then the best model eariler.

#Load a file which has all prediction including of combined model
df_prediction = combine_all_prediction(r'\model_predictions_combined.csv')

#Split Twitter vs Non-Twitter records to to get metrics separately
y_act          = df_prediction['label']

df_prediction1 = df_prediction[df_prediction.twitter=='Y']
y_act1         = df_prediction[df_prediction.twitter=='Y']['label']

df_prediction2 = df_prediction[df_prediction.twitter=='N']
y_act2         = df_prediction[df_prediction.twitter=='N']['label']

results1 = calculate_metrics(y_act, df_prediction)

print ("Number of Models {modnum}".format( modnum = len(results1)))

#Results for all Records
results1 = calculate_metrics(y_act, df_prediction)
results1
results1.sort_values("Accuracy",ascending=False)

#Save Metrics in a file for Reporting Purpose
fl= resultsfolder + r'\model_results_all.csv'
print ('All Results Prediction Results in file ', fl)
results1 = calculate_metrics(y_act, df_prediction)
results1.to_csv(fl)


f2= resultsfolder + r'\model_results_twitter.csv'
print ('Twitter Statement Prediction Results in file ', fl)
results1 = calculate_metrics(y_act1, df_prediction1)
results1.to_csv(f2)


f3= resultsfolder + r'\model_results_wo_twitter.csv'
print ('Normal Statements Prediction Results in file ',fl)
results1 = calculate_metrics(y_act2, df_prediction2)
results1.to_csv( f3)
